+++ src/index.ts
@@ -1,5 +1,5 @@
 /** @ts-ignore , vue-demi seems to be not strongly typed so typescript freaks out. */
-import { App, Vue2, Plugin, PluginObject, isVue2 } from "vue-demi";
+import { App, Vue2, Plugin, PluginObject } from "vue-demi";
 import {
   SupabaseClient,
   SupabaseClientOptions,
@@ -14,18 +14,7 @@ import {
   useSupabaseStorage,
 } from "./composables";
 
-// @ts-expect-error: Vue2 types for a Vue# project
-declare module "vue/types/vue" {
-  interface Vue {
-    $supabase: SupabaseClient;
-  }
-}
-
-type Options = {
-  supabaseUrl: string;
-  supabaseKey: string;
-  supabaseOptions: SupabaseClientOptions;
-};
+import { Options } from "../vue-supabase";
 
 function install(app: typeof Vue2 | App, options: Options) {
   const supabase = new VueSupabaseClient(
